requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.8.0
djangorestframework-jwt==1.11.0
Markdown==2.6.11
Pillow==5.1.0
pkg-resources==0.0.0
PyJWT==1.6.1
pytz==2018.3
